did stuffs

+++ sns-client/src/main/java/com/github/eeriefoods/snsclient/controller/CourseToolBarController.java
@@ -1,6 +1,7 @@
 package com.github.eeriefoods.snsclient.controller;
 
 import com.github.eeriefoods.snsclient.model.Course;
+import com.github.eeriefoods.snsclient.shared.NotificationHandler;
 
 public class CourseToolBarController extends ToolBar{
     @Override
@@ -9,11 +10,14 @@ public class CourseToolBarController extends ToolBar{
         TFDSearch.setOnKeyTyped(event -> courseSearch());
 
         BTNCreate.setOnAction(event -> mainController.switchBar(mainController.tabPane.getSelectionModel().getSelectedItem()));
+
         BTNDelete.setOnAction(event -> {
-            Course course = courseTableView.getSelectionModel().getSelectedItem();
-            courseTableController.deleteCourse(course);
-            courseTableView.getItems().remove(course);
-        });
+            NotificationHandler nh = new NotificationHandler();
+            if (nh.askForConfirmation("Wirklich löschen?", "Im Kurs " + courseTableView.getSelectionModel().getSelectedItem().getId() + " befinden sich noch " + " Student:innen. Soll der Kurs wirklich gelöscht werden?") == true) {
+                Course course = courseTableView.getSelectionModel().getSelectedItem();
+                courseTableController.deleteCourse(course);
+                courseTableView.getItems().remove(course);
+        }});
 
     }
